# Conflicts:
#	voyageone-static/2.0/bs-config.js
#	voyageone-static/2.0/develop/modules/cms/translate/en.js
#	voyageone-static/2.0/develop/modules/cms/translate/zh.js
#	voyageone-static/2.0/develop/static/css/app.pages.css
#	voyageone-tools/voyageone-dao-codegen/src/main/java/com/voyageone/tools/dao/codegen/VOCmsDaoGenerator.java

+++ voyageone-service2/voyageone-service2-model/src/main/java/com/voyageone/service/bean/cms/jumei/BatchUpdatePriceParameterBean.java
@@ -9,7 +9,7 @@ public class BatchUpdatePriceParameterBean {
     int priceType;
     double discount;
     double price;
-    List<Long> listPromotionProductId;
+    List<Integer> listPromotionProductId;
     int jmPromotionId;
 
     public int getJmPromotionId() {
@@ -52,11 +52,11 @@ public class BatchUpdatePriceParameterBean {
         this.price = price;
     }
 
-    public List<Long> getListPromotionProductId() {
+    public List<Integer> getListPromotionProductId() {
         return listPromotionProductId;
     }
 
-    public void setListPromotionProductId(List<Long> listPromotionProductId) {
+    public void setListPromotionProductId(List<Integer> listPromotionProductId) {
         this.listPromotionProductId = listPromotionProductId;
     }
 }
